Create db.sql
Our database. It contains "users", "categories", "postings" and "tags".
Where posts has a foreign key "user_id" which says which user is the author and "cat_id" which says from which category is current post.
Tags has a foreign key "post_id" which says from which post is the tag.